# MM-21211: Optimize Channelstore.UpdateLastViewedAt