@Pomyk 2019-12-27
Did you test performance on channels with high user count? I see that the CTE version uses index idx_channelmembers_channel_id instead of primary key (at least on postgresql 9.4).