@agnivade 2019-12-28
Thanks for looking into it @Pomyk - Yes I have tested the performance using our loadtest which uses 1000 users which is the standard size of our installations.

From what I am seeing using postgres11, it is using the primary key:
```
 CTE Scan on c  (cost=34.01..34.05 rows=2 width=78) (actual time=0.039..0.039 rows=0 loops=1)
   CTE c
     ->  Bitmap Heap Scan on channels  (cost=8.58..15.65 rows=2 width=43) (actual time=0.037..0.038 rows=0 loops=1)
           Recheck Cond: ((id)::text = ANY ('{sfsg,25235f}'::text[]))
           ->  Bitmap Index Scan on channels_pkey  (cost=0.00..8.58 rows=2 width=0) (actual time=0.034..0.035 rows=0 loops=1)
                 Index Cond: ((id)::text = ANY ('{sfsg,25235f}'::text[]))
   CTE updated
     ->  Update on channelmembers cm  (cost=0.41..18.37 rows=1 width=319) (actual time=0.003..0.003 rows=0 loops=1)
           ->  Nested Loop  (cost=0.41..18.37 rows=1 width=319) (actual time=0.001..0.001 rows=0 loops=1)
                 ->  CTE Scan on c c_1  (cost=0.00..0.04 rows=2 width=196) (actual time=0.001..0.001 rows=0 loops=1)
                 ->  Index Scan using channelmembers_pkey on channelmembers cm  (cost=0.41..8.43 rows=1 width=193) (never executed)
                       Index Cond: (((channelid)::text = (c_1.id)::text) AND ((userid)::text = '2455'::text))
```

Are you saying you are not seeing the userId index being used at all in 9.4 ?