@Pomyk 2019-12-28
Tested on 9.4 and 9.5 and the plan is:
```
 CTE Scan on c  (cost=16.73..16.75 rows=1 width=78) (actual time=0.051..0.052 rows=1 loops=1)
   CTE c
     ->  Index Scan using channels_pkey on channels  (cost=0.28..8.30 rows=1 width=43) (actual time=0.048..0.049 rows=1 loops=1)
           Index Cond: ((id)::text = 'rjemtnjdmpbeuj1ordr98ftbnh'::text)
   CTE updated
     ->  Update on channelmembers cm  (cost=0.29..8.44 rows=1 width=342) (actual time=0.376..0.376 rows=0 loops=1)
           ->  Nested Loop  (cost=0.29..8.44 rows=1 width=342) (actual time=0.060..0.178 rows=1 loops=1)
                 ->  CTE Scan on c c_1  (cost=0.00..0.02 rows=1 width=196) (actual time=0.008..0.008 rows=1 loops=1)
                 ->  Index Scan using idx_channelmembers_channel_id on channelmembers cm  (cost=0.29..8.41 rows=1 width=216) (actual time=0.049..0.166 rows=1 loops=1)
                       Index Cond: ((channelid)::text = (c_1.id)::text)
                       Filter: ((userid)::text = 'r44fm6ncmbn8uqeb4ozjxjd9ho'::text)
                       Rows Removed by Filter: 160
```