@Pomyk 2019-12-28
After analyze it started using primary key.