@agnivade 2019-12-30
There are already existing tests here - testChannelStoreUpdateLastViewedAt. Those test all the fields touched in the method to verify there are no regressions. Did you have any specific check in mind ?